Modify: RoomService 분리
merge 오류 발생가능성으로 인한 분리
